organizes healthchecks a bit better. makes healthcheck module
