+++ lib/healthchecks.rb
@@ -1,0 +1,4 @@
+module Healthchecks
+    require 'healthchecks/queue_dead_set_check'
+    require 'healthchecks/queue_latency_check'
+end
