+++ lib/healthchecks/queue_latency_check.rb
@@ -1,15 +1,6 @@
 # frozen_string_literal: true
 
 module HealthChecks
-  class QueueDeadSetCheck < OkComputer::Check
-    def check
-      ds = Sidekiq::DeadSet.new
-      size = ds.size
-      mark_failure if size.positive?
-      mark_message "There are #{size} messages in the DeadSet Queue"
-    end
-  end
-
   class QueueLatencyCheck < OkComputer::Check
     attr_accessor :threshold
